+++ client/src/PasswordReset.tsx
@@ -4,7 +4,7 @@ import Button from './common/Button';
 import Input from './common/Input';
 import message from './common/message';
 import Spacer from './common/Spacer';
-import { api } from './utilities/fetch-json';
+import { api } from './utilities/api';
 
 function PasswordReset({ passwordResetId }: { passwordResetId: string }) {
   const [email, setEmail] = useState('');
