+++ client/src/SignIn.tsx
@@ -6,7 +6,7 @@ import ButtonLink from './common/ButtonLink';
 import Input from './common/Input';
 import message from './common/message';
 import Spacer from './common/Spacer';
-import { api } from './utilities/fetch-json';
+import { api } from './utilities/api';
 import useAppContext from './utilities/use-app-context';
 
 function SignIn() {
