+++ client/src/app-header/AppMenu.tsx
@@ -3,7 +3,7 @@ import DotsVerticalIcon from 'mdi-react/DotsVerticalIcon';
 import React, { useState } from 'react';
 import { Redirect } from 'react-router-dom';
 import IconMenu from '../common/IconMenu';
-import { api } from '../utilities/fetch-json';
+import { api } from '../utilities/api';
 import useAppContext from '../utilities/use-app-context';
 import AboutModal from './AboutModal';
