+++ client/src/queryEditor/toolbar/QueryTagsModal.tsx
@@ -3,7 +3,7 @@ import Modal from '../../common/Modal';
 import MultiSelect from '../../common/MultiSelect';
 import { setQueryState } from '../../stores/editor-actions';
 import { useEditorStore } from '../../stores/editor-store';
-import { api } from '../../utilities/fetch-json';
+import { api } from '../../utilities/api';
 
 function QueryTagsModal({ visible, onClose }: any) {
   const tags = useEditorStore<string[]>((s) => s?.query?.tags || []);
