+++ client/src/utilities/runQueryViaBatch.ts
@@ -1,4 +1,4 @@
-import { api } from './fetch-json';
+import { api } from './api';
 
 function sleep(ms: any) {
   return new Promise((resolve) => setTimeout(resolve, ms));
