+++ client/src/utilities/swr-fetcher.ts
@@ -1,4 +1,4 @@
-import { api } from './fetch-json';
+import { api } from './api';
 
 export default async function swrFetcher(url: any) {
   const { data, error } = await api.get(url);
