+++ client/src/utilities/use-app-context.ts
@@ -1,5 +1,5 @@
 import baseUrl from './baseUrl';
-import { api } from './fetch-json';
+import { api } from './api';
 
 function useAppContext() {
   let { data } = api.useAppInfo();
